super wip

+++ src/ray.rs
@@ -3,7 +3,6 @@ use na::Point3;
 use na::Vector3;
 
 pub struct Ray {
-    pub origin: Point3<f32>,
-    pub direction: Vector3<f32>,
+    pub origin: Point3<f64>,
+    pub direction: Vector3<f64>,
 }
-
